+++ socks5.c
@@ -16,7 +16,7 @@
 #include "res.h"
 #include "tcp.h"
 
-int socks5connect(struct addrinfo *ai, int timeout, const char *socks5_username, const char *socks5_password, const char *host, int port, char abort_on_resolve_failure)
+int socks5connect(struct addrinfo *ai, double timeout, const char *socks5_username, const char *socks5_password, const char *host, int port, char abort_on_resolve_failure)
 {
 	struct sockaddr_in sai;
 	uint32_t addr = 0;
